requirements.txt
python-telegram-bot
transformers
diffusers
moviepy
torch